Product submission Complete

+++ database/migrations/2022_03_30_125837_create_product_materials_table.php
@@ -24,7 +24,7 @@ class CreateProductMaterialsTable extends Migration
                     ->references('id')
                     ->on('materials')
                     ->onDelete('cascade');
-            $table->integer('stock');       
+            $table->float('stock');       
         });
     }
